requirements.txt
Django==1.11.*
django-jalali-date==0.3.*
mysqlclient
pytz==2021.1
persiantools==1.5.*
zeep==3.4.*
requests==2.25.*
django-ratelimit==2.0.*
django-heartbeat==2.0.*